backend/HouseTour/requirements.txt
openai
python-dotenv
pillow
gtts
ipython
langchain
langchain-community
langchain-openai
faiss-cpu
sentence-transformers
tiktoken
pyaudio
groq
elevenlabs
langchain-groq
numpy
pydub



